requirements.txt
ops >= 1.2.0
jinja2
ops-lib-pgsql
GitPython
